+++ src/src/store.c
@@ -97,8 +97,8 @@ and down as store is reset or released. nonpool_malloc is the total got by
 malloc from other calls; this doesn't go down because it is just freed by
 pointer. */
 
-static int pool_malloc = 0;
-static int nonpool_malloc = 0;
+static ssize_t pool_malloc = 0;
+static ssize_t nonpool_malloc = 0;
 
 /* This variable is set by store_get() to its yield, and by store_reset() to
 NULL. This enables string_cat() to optimize its store handling for very long
@@ -126,7 +126,7 @@ Returns:      pointer to store (panic on malloc failure)
 */
 
 void *
-store_get_3(int size, const char *filename, int linenumber)
+store_get_3(size_t size, const char *filename, int linenumber)
 {
 /* Round up the size to a multiple of the alignment. Although this looks a
 messy statement, because "alignment" is a constant expression, the compiler can
@@ -142,8 +142,8 @@ these functions are mostly called for small amounts of store. */
 
 if (size > yield_length[store_pool])
   {
-  int length = (size <= STORE_BLOCK_SIZE)? STORE_BLOCK_SIZE : size;
-  int mlength = length + ALIGNED_SIZEOF_STOREBLOCK;
+  size_t length = (size <= STORE_BLOCK_SIZE)? STORE_BLOCK_SIZE : size;
+  size_t mlength = length + ALIGNED_SIZEOF_STOREBLOCK;
   storeblock *newblock = NULL;
 
   /* Sometimes store_reset() may leave a block for us; check if we can use it */
@@ -195,10 +195,10 @@ linenumber = linenumber;
 DEBUG(D_memory)
   {
   if (running_in_test_harness)
-    debug_printf("---%d Get %5d\n", store_pool, size);
+    debug_printf("---%d Get %5u\n", store_pool, (unsigned int)size);
   else
-    debug_printf("---%d Get %6p %5d %-14s %4d\n", store_pool,
-      store_last_get[store_pool], size, filename, linenumber);
+    debug_printf("---%d Get %6p %5u %-14s %4d\n", store_pool,
+      store_last_get[store_pool], (unsigned int)size, filename, linenumber);
   }
 #endif  /* COMPILE_UTILITY */
 
@@ -229,7 +229,7 @@ Returns:      pointer to store (panic on malloc failure)
 */
 
 void *
-store_get_perm_3(int size, const char *filename, int linenumber)
+store_get_perm_3(size_t size, const char *filename, int linenumber)
 {
 void *yield;
 int old_pool = store_pool;
@@ -328,7 +328,7 @@ store_reset_3(void *ptr, const char *filename, int linenumber)
 storeblock *bb;
 storeblock *b = current_block[store_pool];
 char *bc = (char *)b + ALIGNED_SIZEOF_STOREBLOCK;
-int newlength;
+size_t newlength;
 
 /* Last store operation was not a get */
 
@@ -492,15 +492,15 @@ Returns:      pointer to gotten store (panic on failure)
 */
 
 void *
-store_malloc_3(int size, const char *filename, int linenumber)
+store_malloc_3(size_t size, const char *filename, int linenumber)
 {
 void *yield;
 
 if (size < 16) size = 16;
-yield = malloc((size_t)size);
+yield = malloc(size);
 
 if (yield == NULL)
-  log_write(0, LOG_MAIN|LOG_PANIC_DIE, "failed to malloc %d bytes of memory: "
+  log_write(0, LOG_MAIN|LOG_PANIC_DIE, "failed to malloc %zd bytes of memory: "
     "called from line %d of %s", size, linenumber, filename);
 
 nonpool_malloc += size;
@@ -519,12 +519,12 @@ is not filled with zeros so as to catch problems. */
 if (running_in_test_harness)
   {
   memset(yield, 0xF0, (size_t)size);
-  DEBUG(D_memory) debug_printf("--Malloc %5d %d %d\n", size, pool_malloc,
+  DEBUG(D_memory) debug_printf("--Malloc %5zd %d %d\n", size, pool_malloc,
     nonpool_malloc);
   }
 else
   {
-  DEBUG(D_memory) debug_printf("--Malloc %6p %5d %-14s %4d %d %d\n", yield,
+  DEBUG(D_memory) debug_printf("--Malloc %6p %5zd %-14s %4d %d %d\n", yield,
     size, filename, linenumber, pool_malloc, nonpool_malloc);
   }
 #endif  /* COMPILE_UTILITY */
